refact: fill temp file
